requirements.txt
pandas
scikit-learn
torch
torch_geometric
rdkit-pypi
tqdm
pyyaml
tensorboard